Move constant evaluation methods to a separate source file
